Make a package/project out of ICNBenchmarks
