+++ xp/Project.toml
@@ -1,4 +1,5 @@
-name = "ICN benchmarks"
+name = "ICNBenchmarks"
+uuid = "7d4a60b7-6f97-4d93-8ece-c2667b1c39e3"
 authors = ["Jean-François BAFFIER", "Khalil CHRIT", "Florian RICHOUX", "Pedro PATINHO", "Salvador ABREU"]
 
 [deps]
@@ -11,5 +12,11 @@ DrWatson = "634d3b9d-ee7a-5ddf-bec9-22491ea816e1"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
-DrWatson = "2.4.0"
-julia = "1.7.0"
+BenchmarkTools = "1"
+CSV = "0.9"
+CompositionalNetworks = "0.2"
+ConstraintDomains = "0.2"
+Constraints = "0.2"
+DrWatson = "2"
+Tables = "1"
+julia = "1.7"
